+++ Changes.md
@@ -12,6 +12,7 @@ you spot any mistakes.
 * Fix early detection of bad callback to `connection.query`
 * Support Node.js 12.x #2211
 * Support Node.js 13.x
+* Support non-enumerable properties in object argument to `connection.query` #2253
 * Update `bignumber.js` to 9.0.0
 * Update `readable-stream` to 2.3.7